Modularized the project to make it easy to work with others
