+++ game.py
@@ -1,5 +1,7 @@
 import pygame, sys
-import math
+from game_functions import *
+from game_classes import *
+from properties import *
 
 # Need to implement: 
 #
@@ -51,123 +53,19 @@ import math
 #***************************************************************
 
 
-class Player(object): 
-
-    def __init__(self, body, colour, screen_width, screen_height, x, y): 
-        self.body = body
-        self.colour = colour
-        self.screen_width = screen_width
-        self.screen_height = screen_height  
-        self.horizontal_speed = x
-        self.vertical_speed = y
-
-    def update(self, x = 0, y = 0):
-
-        # Make sure the player can't leave the screen. 
-
-        if ((self.body.right >= self.screen_width and x > 0) or 
-            (self.body.x <= 0 and x < 0)):
-
-            self.body = self.body.move(0, y)
-        elif ((self.body.bottom >= self.screen_height and y > 0) or 
-            (self.body.y <= 0 and y < 0)):
-
-            self.body = self.body.move(x, 0)
-        else:
-            self.body = self.body.move(x, y)
-
-
-class Enemy(object):
-
-        # Add health 
-        def __init__(self, body, colour, speed, screen_width, screen_height, hp = 100):
-
-            self.body = body 
-            self.speed = speed 
-            self.colour = colour 
-            self.screen_width = screen_width
-            self.screen_height = screen_height
-            self.hp = hp
-
-
-        def update(self):
-            self.body = self.body.move(self.speed)
-
-        # Return true if the enemy has left the screen 
-        def check_destroy(self):
-            if self.body.right >= self.screen_width or self.hp <= 0:
-                return True
-            else:
-                return False
-
-
-class Bullet(object):
-
-    def __init__(self, speed, pos, colour, screen_height, width = 5, height = 20):
-        self.speed = speed
-        # Inital position of the bullet ie. where the player made the shot 
-        self.body = pygame.Rect(pos, (width, height))
-        self.colour = colour 
-        self.screen_height = screen_height 
-
-    def update(self):
-        self.body = self.body.move(self.speed) 
-
-    # Return true if the bullet leaves the screen 
-    def check_destroy(self):
-        if self.body.bottom >= self.screen_height:
-            return True
-        else:
-            return False
-
-    # If the bullet hits an enemy, remove the enemy and return true
-    # otherwise return false 
-    # def check_hit(self, enemy_list):
-    #   for enemy in enemy_list:
-    #       if self.body.colliderect(enemy.body):
-    #           enemy_list.remove(enemy)
-    #           print "hit"
-    #           return 1
-    #       else:
-    #           return 0
-
-class Tower(object):
-
-    def __init__(self, pos, colour, max_range, damage, cost, size):
-        space_between = 40
-        self.type = colour # Each tower type will have a different colour  
-        self.max_range = max_range  # Range is the radius 
-        self.damage = damage 
-        self.cost = cost 
-        self.body = pygame.Rect(pos, size)
-        self.outer_body = pygame.Rect((pos[0] - self.body.width, pos[1] - self.body.height),
-            (self.body.width + space_between, self.body.height + space_between))
-
-BLUE = (0, 128, 255)
-RED = (255, 51, 51)
-ORANGE = (255, 153, 51)
-GREEN = (0, 255, 0)
-BLACK = (0, 0, 0)
-frame_rate = 30 
-tick_speed = 300 
 pygame.init() 
-
-width = 1024
-height = 768
-size = (width, height) 
-screen = pygame.display.set_mode(size) 
-
+screen = pygame.display.set_mode(SIZE) 
 clock = pygame.time.Clock()
 
-player = Player(pygame.Rect((10, 10), (30, 30)), BLUE, width, height, 
-    tick_speed / frame_rate, tick_speed / frame_rate)
+player = Player(pygame.Rect((10, 10), (30, 30)), BLUE, WIDTH, HEIGHT, 
+    TICK_SPEED / FRAME_RATE, TICK_SPEED / FRAME_RATE)
 
 # enemy stuff
 enemy_list = [] 
 counter = 0
 spawn_rate = 50
 enemy_size = (30, 30)
-enemy_start = (10, height / 2)
+enemy_start = (10, HEIGHT / 2)
 
 # Direction constants
 RATE = 3 
@@ -179,10 +77,10 @@ enemy_speed = RIGHT
 
 # Enemy pathing
 turn_size = (45, 45)
-enemy_turn_list = [(pygame.Rect((width / 4, height / 2), turn_size), UP), 
-                   (pygame.Rect((width / 4, height / 8), turn_size), RIGHT),
-                   (pygame.Rect((width - width / 4, height / 8), turn_size), DOWN), 
-                   (pygame.Rect((width - width / 4, height - height / 4), turn_size), RIGHT)]
+enemy_turn_list = [(pygame.Rect((WIDTH / 4, HEIGHT / 2), turn_size), UP), 
+                   (pygame.Rect((WIDTH / 4, HEIGHT / 8), turn_size), RIGHT),
+                   (pygame.Rect((WIDTH - WIDTH / 4, HEIGHT / 8), turn_size), DOWN), 
+                   (pygame.Rect((WIDTH - WIDTH / 4, HEIGHT - HEIGHT / 4), turn_size), RIGHT)]
 
 # Bullet stuff 
 bullet_list = [] 
@@ -199,9 +97,6 @@ tower_size = {GREEN : (20, 20)}
 tower_range = {GREEN : 160}
 
 # Functions
-def distance (p1, p2):
-    return math.sqrt((p1[0] - p2[0])**2 +(p1[1] - p2[1])**2)
-
 while True:
     for event in pygame.event.get():
         if event.type == pygame.QUIT: 
@@ -211,7 +106,7 @@ while True:
     # ticks have passed an enemy will spawn. 
     if counter % spawn_rate == 0:
         enemy_list.append(Enemy(pygame.Rect(enemy_start, enemy_size), RED,
-                                enemy_speed, width, height))
+                                enemy_speed, WIDTH, HEIGHT))
 
     for enemy in enemy_list:
         for turn in enemy_turn_list:
@@ -231,7 +126,7 @@ while True:
     if pressed[pygame.K_SPACE]:
         now = pygame.time.get_ticks()
         if now - last_shot >= SHOT_DELAY:
-            bullet_list.append(Bullet(bullet_speed,((2 * player.body.x + player.body.width)/2 - 5/2, player.body.y), BLUE, height))
+            bullet_list.append(Bullet(bullet_speed,((2 * player.body.x + player.body.width)/2 - 5/2, player.body.y), BLUE, HEIGHT))
             last_shot = now
 
     if pressed[pygame.K_t]:
@@ -308,5 +203,5 @@ while True:
         enemy.update()
 
     pygame.display.flip()
-    clock.tick(frame_rate)
+    clock.tick(FRAME_RATE)
     counter = counter + 1
