+++ game_functions.py
@@ -1,0 +1,6 @@
+import math
+
+# A module containing functions in the game
+def distance (p1, p2):
+    return math.sqrt((p1[0] - p2[0])**2 +(p1[1] - p2[1])**2)
+
