+++ properties.py
@@ -1,0 +1,15 @@
+# Game properties can be changed/added here
+
+#Colours
+BLUE = (0, 128, 255)
+RED = (255, 51, 51)
+ORANGE = (255, 153, 51)
+GREEN = (0, 255, 0)
+BLACK = (0, 0, 0)
+FRAME_RATE = 30 
+TICK_SPEED = 300 
+
+# Window size properties
+WIDTH = 1024
+HEIGHT = 768
+SIZE = (WIDTH, HEIGHT)
